Migrate to v2 particle library format, change author and bump v2.1.1
