+++ library.properties
@@ -1,0 +1,7 @@
+name=M2XStreamClient
+version=2.1.1
+license=MIT
+author=AT&T M2X <attm2x-internal@citrusbyte.com>
+sentence=Client Library for AT&T's M2X - Cloud-Based Data Storage Service and Management Toolset Customized for the Internet of Things
+url=https://github.com/attm2x/m2x-particle-core
+repository=https://github.com/attm2x/m2x-particle-core.git
