+++ src/M2XStreamClient/M2XStreamClient.h
@@ -1,0 +1,1 @@
+#include "../M2XStreamClient.h"
