+++ src/M2XStreamClient_template.h
