+++ src/StreamParseFunctions.h
